PySide6: Add QKey support for the setShortcut method
When this method is used in Qt/C++, the argument gets internally cast
to a QKeyCombination if it is a QKey, but in Python land this is not
automatically covered.

A workaround was for the users to manually cast the QKey, but this
adds an extra step to achieve the same.

This patch adds a new method to QAction and QAbstractButton to use
setShortcut with a QKey as an argument.

Pick-to: 6.1
Change-Id: I9e9ebe16a65cb5aeb26a367faecbbbd414d80e03
Fixes: PYSIDE-907
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>