PySide6: Add vertexDataAsPoint2D()/setVertexDataAsPoint2D()
vertexDataAsPoint2D returns a pointer (const/non-const for
modification) to a series of Point2D within a geometry, but since in
Python we don't have access to it, this change replaces that signature
for one that returns a list of Point2D elements.

For Pythonicity, also add a setter setVertexDataAsPoint2D() to enable
modification.

Fixes: PYSIDE-1345
Change-Id: I6aba2a097022637c398ad94c17f342f13b4da449
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>