remove QBackingStore from qrasterwindow_test.py
The QBackingStore part of qrasterwindow_test.py does not work
on macOS. We remove this part of the test that should not
even reside in PySide but is a Qt matter.

This was first done in Qt 6, but is now also needed
in 5.15 since apparently a relevant change in Qt
was cherry-picked.

Change-Id: I520a5fd1336850d3e3883bcccb9b6a4b2cdafeda
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 35f0aa7be658c6dc5887aad6c38e098e3466ec5f)