write a leak_finder script
PySide has a few known memory leaks and probably some
unknown leaks as well.

There are more or less unproblematic pseudo-leaks where
only too many references are kept, but there are also the
true leaks, where new objects are created all the time.

This script is a first attempt to get rid of that problem.
The interface is not perfect but works pretty well.

Change-Id: I8e44eedde036f25adadf66ca2a7fdac8095619c4
Pick-to: 6.0
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>