basewrapper: clear all traces of special Q*Application treatment
After simplifying qApp so much, it is no longer necessary to treat
Q*Application special. We now can allow garbage collection for
all objects which simplifies the code quite much.

Then it was easy to shrink the implementation to a single function
`MakeQAppWrapper`, which made the whole "qapp_macro" files no
longer necessary.

This cosmetic change will probably not be back-ported to 5.15 .

Task-number: PYSIDE-1447
Change-Id: I8ae3c5575e62bd7b5d6bac65c25775c712bb178f
Pick-to: 6.0
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>