allow to use shiboken.delete() to delete Q*Application
This function should initially not allow deleting Q*Application
with a shiboken.delete() function, but unfortunately some
competitor can do that, so we will as well.

Deletion of qApp needs knowledge about qApp.
We add a hook function that calls the pyside function
  `destroyQCoreApplication`.

A special problem was testing interactively.
See the error description for the days long journey :)

Change-Id: I34862425c2cb2cc80d6cafc22d25a867f96f3e0a
Fixes: PYSIDE-1470
Pick-to: 6.0
Pick-to: 5.15
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>