feature: Disable selection while creating a type
PySide 6 suddenly has problems with feature switching during
subtype initialization.

We introduce an enable function that can temporarily suppress
switching. This problem is solved so far.

It is the question whether this will break again in another
constellation. It might be considerable for the future to have
something like Python's PyType_Ready function.
Right now this is too much effort.

Change-Id: If0ed786d4761cf2356f01f7478e4a0d750e88d3c
Fixes: PYSIDE-1463
Pick-to: 6.0
Pick-to: 5.15
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>