shiboken: Provide the correct inheritance for enum types
This works for Python 2, already.

Python 3 has the extra complication of a special long integer
implementation that is hard to get rorrectly mapped to the
enum structure.

There is still some work necessary to let objects look right, although
we are inheriting from `ìnt`.

Development happens on 5.15, because it still has to cope with
both Python versions.
This will finally go into dev .

TODO: Rewrite this, transfer to dev

Change-Id: Ifaaea40a4ddf2337e565fb57e6a69b7cc9a6040f
Fixes: PYSIDE-1347
Pick-to: 6.0
Pick-to: 5.15
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>