shiboken: Provide the correct inheritance for enum types
Originally, it was planned to follow a competitor and derive enums
from int objects. But this was abandoned in favor of Python Enums.

We therefore simply leave most code as it is and only insert a
Shiboken::Enum class that accepts an optional integer.
This class is published by shiboken with signature.

Derived classes still have no signature themselves, but they all
inherit from this known class Enum. This is intentional to avoid
creating useless signatures for every derived Enum.

A test was included.

Change-Id: Ifaaea40a4ddf2337e565fb57e6a69b7cc9a6040f
Fixes: PYSIDE-1347
Pick-to: 6.0
Pick-to: 5.15
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>