move "shiboken6/shiboken6" to "shiboken6/Shiboken"
Modifying the paths to work in the new way is a quite ambitious
task. But doing so improves the overall project structure and
makes imports unambiguous.

This patch should not be applied alone but with
    move "shiboken6/shiboken6" to "shiboken6/Shiboken" temp

The reworked version of this patch no longer has different
structures in `build` and `install`.

Tested with
Python 3.6, 3.7, 3.8, 3.9 debug build
Python 3.6 debug install
Python 3.9 release install

Task-number: PYSIDE-1497
Change-Id: Id9d816dd825907f9359651e7e2f69f54e1ba46c9
Reviewed-by: Christian Tismer <tismer@stackless.com>