shiboken: Provide the correct inheritance for enum types, Fix
This patch adds a missing Py_INCREF on the base class
`Shiboken.Enum` when a new type is created. This is _not_
automatically provided (but it _is_  for the meta type...).

The error was visible in a debug build, only.

Change-Id: I213713535acd1dde71bd8e3a0efe6122119d43b5
Fixes: PYSIDE-1487
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit ddb03b40f311a9c34b599136ac1cb0324f6b2b42)
Reviewed-by: Christian Tismer <tismer@stackless.com>