PySide: write a renamer script for debugging, improved
Debug output is hard to read if there are many objects with the
same type but different address.

This script builds simple names instead of addresses which are easy
to track. See the info at the beginning of the script.

The improved version is more universal since it only requires
a "0x..." word followed by some name. There is no other fixed formatting.

Task-number: PYSIDE-79
Task-number: PYSIDE-1478
Change-Id: Iae4abd447dc14a67fe7cc212c1de337c70499ac1
Pick-to: 6.0
Pick-to: 5.15
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>