feature: Supply switching for newly created objects
Switching was quite complete, but when the module with
a feature selection created a new object, no switching
of the new object happens.

This was a pretty hard to find bug, but the solution was trivial.

Fixes: PYSIDE-1478
Change-Id: I663a562d4d9512b627b3add37df0908f9d785e1f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit d158720ec4dd3fe8623faaf9afcb2e7f9f368e51)
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>