shiboken6: Fix inline wrapper being generated for the wrong class
When generating the inline wrappers in the header, the implementing
class should be used instead of the owner, otherwise things can
go wrong in case of multiple inheritance (for example
function setProperty() being inherited from QObject and
QDesignerPropertySheetExtension).

Since the code writing such a wrapper is duplicated
in HeaderGenerator, move it to a separate function.

Task-number: PYSIDE-1455
Change-Id: I9481d0d64802ac2c9251759958fb8345863d6889
Reviewed-by: Christian Tismer <tismer@stackless.com>