shiboken6: Enable building with --avoid-protected-hack for gcc, clang
Disable the C++ wrapper generation for classes with
private destructors for these compilers.
Fix one additional condition overlooked by
1d044f467070a040713c9566a8a8de3a56c571e7.
With that, it is in principle possible to build
with g++ and --avoid-protected-hack except for
the PrivateDtor shiboken test, which tests accessing
protected methods for classes with private destructors.

Task-number: PYSIDE-1202
Task-number: PYSIDE-504
Change-Id: Ie40ad56c913dd544303aa7d795e80282afb43705
Reviewed-by: Christian Tismer <tismer@stackless.com>