examples: directory structure
When generating the gallery automatically, the structure
of the examples was not the same for all the modules.

This reorganization was made under the following assumption:

An example requires:
- its own directory,
- a '.pyproject' file,
- one or more '.py' files,
- extra files (likes images, .rc, .ui, etc)

Even for cases when  subdirectories are made to depict categories, like
'widgets/state-machine/...', one directory is required per example.

There should not be '.pyproject' files for categories,
except for the general 'examples.pyproject' to open all of them.

Task-number: PYSIDE-1490
Pick-to: 6.0
Change-Id: Ie690773c816a248016dbebaad90fbb1d1c2aeea5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>