PySide6: Add a Designer plugin
Add a convencience class QPyDesignerCustomWidgetCollection to
the Qt Designer module, which provides functions for
registering widget types or adding QDesignerCustomWidgetInterface
instances. A static instance of it is stored as a dynamic
property on QCoreApplication, which is retrieved by a Qt Designer
plugin, which provides the collection of widgets registered in
Python.

Task-number: PYSIDE-1455
Change-Id: If4055e6c9db6a03b32016b013a1130051bbd472a
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>