Switch from os.path to pathlib.Path, systematic test changes
This patch does the systematic changes on all tests.

Note that there was a bug in principle:

    os.path was used without importing os.path

This worked through the way how the os module initializes
itself, but you always will find the explicit import in
the python library.

But the problem is going away anyway after the transition
to pathlib.Path :)

Change-Id: I95144d7f9bb07a38376c3aa428df663f2e64bcb7
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>