doc: add tool to generate examples gallery
This script generates a gallery for all the example
currently in pyside-setup/examples.

Using this tool will overwrite the index rst file for the examples
located in 'sources/pyside6/doc/examples/index.rst

Additionally, to display the code of each example, this will generate
one extra .rst file for each example that contains a .pyproject file,
for example: 'sources/pysides6/doc/examples/example_widgets__tetrix.rst'

Currently, the usage of this tool is not incorporated in the
documentation building process.

Task-number: PYSIDE-1490
Pick-to: 6.0
Change-Id: I78546d4c7905fd8b521f4112457980b4d1d56860
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>