shiboken6/Documentation: Fix the extra sections
The file filter used for filtering the extra documents did not
include the '.', so, the module description QtXmlPatterns.rst
was added as an extra document for QtXml. Add the dot to the
filter and sort the result. Use QDir::entryInfoList() since
it is faster and gives the full path.

Pick-to: 6.0
Task-number: PYSIDE-841
Change-Id: I173979b9a527121b95bcb0190f603c02565bc282
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>