shiboken6/Documentation: Fix the extra sections
The file filter used for filtering the extra documents did not
include the '.', so, the module description QtXmlPatterns.rst
was added as an extra document for QtXml. Add the dot to the
filter and sort the result. Use QDir::entryInfoList() since
it is faster and gives the full path.

Task-number: PYSIDE-841
Change-Id: I173979b9a527121b95bcb0190f603c02565bc282
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 2e43fc4cfd997de48eae352a8a42936f6e4b3785)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>