signature: Provide fixes and improvements, Part 1, addendum
This patch had a way too complicated method to find
out if something is a method, and this broke on
Python 2 because we don't have the __qualname__
attribute everywhere.

But this can be done much easier: We check the type and see
if it is a function or builtin function.
Everything else must be a method, no matter what it is :)

Change-Id: I07e5690f36e4ed4d50ea5e203233b369dc70fa5f
Task-number: PYSIDE-510
Pick-to: 6.0
Pick-to: 5.15
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>