cppgenerator: simplify code further using AutoDecref
Shiboken::AutoDecref is an increadible handy tool to simplify
reference counted code.

One missing spot was to be able to declare an AutoDecref variable
without specifying a value for later use by reset().
A workaround was this code:

    static PyObject *_dummy{};
    Shiboken::AutoDecref errCode(_dummy);
    ...
    errCode.reset(someVar);

We now allow AutoDecref without an expression and can write

    Shiboken::AutoDecref errCode{};
    ...
    errCode.reset(someVar);

That construct is used to further simplify the generated code quite much.

Change-Id: I2ae8f2214c0fbaae0a935aa0b8d69b4a7922a6c8
Task-number: PYSIDE-1499
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>