doc: use 'doc' directory for examples gallery
The initial patch that generates a gallery based on the examples in the
respository, was limited to only showing the content of the files
in each '.pyproject' file.

That approach didn't allow to describe each example, nor add
complementary images, like screenshots to each example page.

This patch introduces the option to consider everything inside a 'doc/'
directory on the example directory. The files that are copied over are
not directory, but only files, for images to be the main focus.

For example, currently the Tetrix case contained the following files:

    $ ls examples/widgets/tetrix/
    tetrix.py
    tetrix.pyproject

On this patch you can see that now there is a doc directory with the
following content:

    $ ls examples/widgets/tetrix/doc
    tetrix-screenshot.png
    tetrix.rst

The example page that will be generated for this case will contain the
content of the 'doc/tetrix.rst' file, plus the content of all the
project files at the end.

The 'doc/tetrix.rst' file contains a reference to the
'tetrix-screenshot.png' image, so that file will be copied over too.

Task-number: PYSIDE-1112
Pick-to: 6.0
Change-Id: I2d11833c461d8073d2d2888576d876d3f834103a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>