feature-import: Hide confusing feature frame entry
The feature construct adds an extra level of indirection to the
import command that shows up in tracebacks. This is quite
irritating.

To avoid these extra layer and much effort fo fix this in
Python, it is more straight-forward to write a C function and

- calls our redirection and returns if ok or error
- calls the original import when feature not applicable.

The change to the import function is made transparent by the
builtin variables __orig_import__ and __feature_import__.

Task-number: PYSIDE-1338
Task-number: PYSIDE-1019
Change-Id: Iacae6fbc47cf85f366f59ec3f6c5cca66d5ad65e
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>