__feature__: Make imports more safe against installation errors
PySide is sometimes installed by modified installer scripts.

The __feature__ script may be problematic due to dunder usage.
We therefore remove dunder and use it internally, only.

The __feature__ also has a problem with PyInstaller (maybe solved
by the above) and does not work. But if somebody did an omission,
a false error message can be generated.

As a cure, we explicitly test the import of PySide6.support before
we import the pseudo module PySide6.support.__feature__ . This is
an unconditional error, while a missing feature import is just
a warning.

Task-number: PYSIDE-1502
Change-Id: Ia4feaa67b615581291d1c7ff0c5fbdf40a2f176f
Pick-to: 5.15
Reviewed-by: Christian Tismer <tismer@stackless.com>