move "shiboken6/shiboken6" to "shiboken6/Shiboken" temp, fix
The temporary copy from Shiboken.X to shiboken6.X was ok
for the build directory, but the direct copy to install sometimes
fails due to permission problems.

The problem was that it is not easy to know what X is. It can
be abi3.so, or pyd or similar. The workaround was to use
a directory install with a "shiboken6.*" pattern.

A test case is included.
The special handling will be automatically disabled with version
6.2 . The code can but need not be removed after this.

Change-Id: I582373cbc1746f7754a8222815c56ad5608e72da
Fixes: PYSIDE-1514
Task-number: PYSIDE-1497
Reviewed-by: Christian Tismer <tismer@stackless.com>