PySide import: Make imports more safe against installation errors
This patch originally was
    "__feature__: Make imports more safe against installation errors"

After applying the patch to 5.15 with errors, it became clear that
not __feature__ was a problem, but the "import PySideX.support"
did not work in the embedding case.

This observation made the error message concerning PyInstaller
pointless. Instead, the pseudo-import of PySideX.support was fixed
to work correctly in all cases.

Tested on 5.15 with PyInstaller, this is now going back into dev.
Manually cherry-picked to 6.0

Pick-to: 5.15
Task-number: PYSIDE-1502
Change-Id: I2fd680a6a28de00f5392147b0d050deb7663dc22
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>