doc: enable snippets translate tool
Create snippets directories in case they don't exist.
Add "-q" option as the examples directory to keep the "--quiet"
option from the build.

Something to consider with this patch:
Removing all the previous old snippets, which include not only 'cpp'
files, but also '.h', '.qdoc', '.qrc', '.png', and '.qml', which might
be used in the snippets, but will need to be updated in the generation
tool, to not translate, and copy them over.

Pick-to: 6.0
Change-Id: Icbef9e1c93a12b90dbcfa990ef055ca6f8868407
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>