cmake: add dependencies to rebuild libshiboken embedded files
It was often a pain to handle embedded files correctly. Often, when
embedding was used, one could not see any effect after
re-compiling when --reuse-build was specified.
This rendered the --reuse-build option often quite useless.

The new approach adds dependencies for all involved Python
files, and at the same time reduces them to a single list.

Fixes: PYSIDE-1536
Pick-to: 6.0
Pick-to: 5.15
Change-Id: I6ef36c9ffccd2cd1ac290bb8e879b050f633eaf4
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>