Rewrite the classwizard example
The classwizard created some outdated C++ header and source which is
not useful for Qt for Python.

Rewrite it to generate a Python class and add a special page allowing
for specifying properties and signals of QObjects. Add an overwrite
check and a 'Launch' checkbox to the conclusion page.
Use QFormLayout instead QGridLayout for the pages.

Task-number: PYSIDE-1112
Change-Id: Ice158553571e30ea069ceda8873bf165dc704afc
Reviewed-by: Christian Tismer <tismer@stackless.com>