tools: update missing_binding
Updating the tool to be focused on Qt 6.x.
Moving the long structures to a config file,
moving around functions and formating the code.

Also, moving the script to a separate directory
to be able to have a 'requirements.txt' file which
can fetch all the required packages for the comparison

Pick-to: 6.0
Change-Id: I7a52c33a609d0faff01c20977aa1509ebe6dc058
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>