Brush up the mandelbrot example
Port qtbase/0e69349f6f8e9445877ea5b2105973115ad79cf7:

The example refines the image by running a number of passes
with increasing number of iterations, which is not really
visible to the user. Set an informational text string on
the generated image which provides this information
along with the elapsed time.
The idea is to do the same to the corresponding
Qt for Python example to have some sort of speed comparison
for number crunching.

Add a command line option for the number of passes.

Make the window a bit larger to accommodate the
information.

Task-number: PYSIDE-1112
Change-Id: I9313f860947386123166f794b80b541bc057be2c
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>