tools: display screenshot in example gallery
Adding the option to display a screenshot on the
example gallery.

First, the script will look for an image with the same name
as the example, otherwise it will just get the first
in alphabetical order.

Fixes: PYSIDE-1490
Change-Id: Ia0328c84206a3d66854197abe5ecad8d1d28abc5
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit d5dcdf2de980ef252aabaa56d5a9465453a1c9b2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>