PyPySide: check in the rest of PyPy changes
The PyPy branch has been apart from the main branch for a
long time. Meanwhile, almost everything has settled and can
be merged into the main branch.

This has the advantage of a single branch for both
configurations.

This project is still not ready. It will be called ready when
the Mandelbrot example can be run without any changes.
To achieve that, the current threading problem must be solved.

The WIP branch is kept. It now holds only the pypy-blog.txt file.

Task-number: PYSIDE-535
Change-Id: I317b0e921cc6810e1b76c69991a0f8e592135c65
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>