PyPySide: handle QPainter correctly with GC
While testing the examples with PyPy, a number of examples
used QPainter without explicitly calling painter.end() and
crashed.

This works in standard Python, but leaves the painter open
in other implementations, because the implicit deletion of
objects when going out of scope does not work in the
wrapper when garbage collection is used.

Fixed by providing the missing painter.end() calls.

This problem should finally be fixed by making QPainter a
context manager. The same approach was taken by Python.org and
the file open/close functions. The context manager was needed
for implementations like IronPython, Jython and PyPy.

[ChangeLog][PySide6] The examples were adapted to PyPy's
need to close QPainter, explicitly. Eventually, we may turn
this into a context manager.

Change-Id: I18eeeff7df800bafce91a1e5c98c469aa3bcc41b
Pick-to: 6.2
Task-number: PYSIDE-535
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>