PyPySide: Replace type patching, but not by heaptype, 2-5(5)
PyPy does not allow patching of non-heaptypes. We did that
trick for simplicity in the signature module where we created
a `__signature__` attribute for types.

Type patching of `PyType_Type` needed a few things, only, because
our types are already heaptypes.

The plan was originally to also replace the type patching of

    * PyMethodDescr_Type
    * PyCFunction_Type
    * PyStaticMethod_Type
    * PyWrapperDescr_Type

Actually, it is much easier to stop touching these types at all.
Instead, we use the `get_signature` function in all test cases.

[ChangeLog][shiboken6] Type patching of builtin types is not possible
in PyPy. We use the `get_signature` function, instead for now.

Task-number: PYSIDE-535
Change-Id: I0cb892872a3497ac1fcc67c913d05ba42d75014d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>