PyPySide: Do some preliminary cleanups before the transition
The transition to the future PyPy support needs some
cleanups, first.

The goal is to get rid of all macros which handle
type objects. This works in PySide on Python, but not
with PyPy, which is emulating the whole interface.

The shibokengenerator expression

    if (type->isWrapperType() || type->isEnum() || type->isFlags())
        return QString::fromLatin1("*PepType_SGTP(%1)->converter").arg(cpythonTypeNameExt(type));

builds a very unfortunate union of the three possible macro
type extensions that makes it impossible to do incremental
changes. `PepType_SGTP` is therefore now completely gone.
Finding this problem was a major showstopper.

Furthermore, the usage of the __builtins__ dict object
is not PyPy compatible and replaced by the builtins module.

Task-number: PYSIDE-535
Change-Id: Ic32431b665aae43bb8c9735f69a54f7b3fae8b48
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>