Shiboken: simplify and refactor more after PyTypeObject change
After the move to general PyTypeObject usage, some refactoring
is possible:

In pysidesignal.cpp, the function `getTypeName` is refactored. The
name "type" should be used for PyTypeObject if possible.

In cppgenerator.cpp, a lot of reinterpret_cast calls can be avoided,
also the local variable `pyType` can be used conveniently in
function `writeConverterRegister`.

In basewrapper.cpp, the function `introduceWrapperType` is simplified,
and the part relevant for PyPy is clearly separated.

Task-number: PYSIDE-535
Change-Id: If0fffbad7ee04f92d25111a8620263c8973a7bf0
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>