PyPySide: finish the PyType_FromSpec patch
PySide has been prepared for the inclusion of PyPy by
centralizing all PyType_FromSpec related functions.

It was the original plan to leave sbktypefactory as it
is now, but PyPy has a delay of one or two versions until
the updated type handling will be implemented.

The support for PyType_Modified will come, and it will
even work with our feature switching. But not now!

Therefore, the current slightly hackish solution will
be checked in that solves the timing problems of type
creation by an own re-implementation.

As soon as possible, this version will be reverted.

Task-number: PYSIDE-535
Change-Id: I56330f354edff0453607c1e580a10f620ba10209
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>