PyPySide: Replace the type extension macros by functions
[ChangeLog][PySide6] The type extension macros are replaced by
normal function calls. This is not only a pre-requisite for PyPy
support but also an improvement of type safety.

PyPy cannot stand macro extensions to type objects,
because we are unable to patch the object size during
initialization.

This suggested to re-implement the type extension in
a clean way using functions.

The first idea was to keep the macro version for the
limited API, while the function version would be for PyPy
and no limited API, but the function version is much nicer
and allows further refinements.

The enum/flag alias problem is solved. Unfortunately, this
still does not fix the PYSIDE-229 problem.

Task-number: PYSIDE-535
Change-Id: I39270c6d326cb7d54bd58ceed601f5e1e9b0a8d6
Pick-to: 6.1
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>