signature: Refine the decision heuristics of result tuples
Result tuples do not (yet) get information from XML.
Instead, they fetch their info solely from the embedded
signature strings and information encoded in mapping.py .

When a variable is determined as a result variable, it is
removed from the arg list and normally appended to the
return tuple.

It turned out that a special rule is needed:
A variable that comes first in the variable list should
also be first in the result tuple.

This heuristics should be replaced by more introspection.
Right now, this suffices as a working fix.

Fixes: PYSIDE-1409
Change-Id: Ib65682439e2f678380e0f58a42a36506d0766ddf
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 3506158551488788450498f9b4ffe86375172fea)
Reviewed-by: Christian Tismer <tismer@stackless.com>