doc: fatal error when snippet conversion fails
When the snippets_translate tool fails, the build process
continues normally, falling back to use the original C++
snippets.

This problem can be dangerous, since we will not notice if
something is wrong without checking the full log.
This patch make the build process to fail when the tool
doesn't exit normally.
For builds where the documentation generation is skipped,
this change will have no effect.

Complementary to this, the patch includes a new Python file
to determine if the C++ include have an import
counterpart in Python.

Fixes: PYSIDE-1527
Pick-to: 6.0.4
Change-Id: I725b623f60dbc540a6e7834502300c39cd22b5a3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>