create_changelog: add option to exclude pick-to
This enables the option '-e/--exclude' to skip all the
sha1 related to commits that have been already picked to
another previous version.

The motivation comes to be able to generate the
changelog for 6.1, which comes from the dev branch
that already have many patches that are already cherry picked
to previous releases.

The logic is the following:
- A: Get sha1 from the latest previous non-minor release to the new
     minor-release: origin/6.0.3 and origin/6.1 (created)
- B: Get all the commits from there that have a 'Pick-to' entry
- Use the remaining sha1 from: A - B

Change-Id: I1e1fb5c225c13180dc15ea3e92ce4a641f6455f2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>