py3.10-prep: Fix parser.py for changed typing module
The typing module has subtle changes that are not even
documented: Typing types now have a __name__ attribute.
That confused the parser of the pyi generator because
suddenly stingizing

    Callable[..., Optional[str]]

resulted in

    Callable[..., Optional]

because of special rules that return the generic name
of a typing type, which was very unexpected. Finding this bug
took a lot of debugging of the recursive `_resolve_type`
function.

We finally move the debugging_aid string as a function into
lib/tool.py, because this was very helpful.

[ChangeLog][shiboken6] The parser for .pyi files needed an update
because of undocumented changes in typing.py for Python 3.10 .

Task-number: PYSIDE-1436
Change-Id: I3b8f2c1aa52a23014a8a915a0c677af96dfc536f
Reviewed-by: Christian Tismer <tismer@stackless.com>