py3.10-prep: reset the type cache after feature switching
[ChangeLog][PySide6] Feature switching needs to reset the
internal type cache. This was an omittion that becomes
relevant in Python 3.10 .

When using feature switching in Python 3.10, there were
funny effects where switched and un-switched versions appeared
to co-exist.

It turned out that we were hit by function caching that
is now implemented for the LOAD_ATTR opcode.

It was not known that caching would happen at all for
PySide classes because we don't use Py_TPFLAGS_VALID_VERSION_TAG.
But actually, this flag is used internally by Python to do some
optimizations, and we just have to notify the interpreter of
type changes by PyType_Modified().

Task-number: PYSIDE-1436
Change-Id: Ie8a73f62bd6e9b8156b8ea23626fabd44700158b
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>