py3.10-prep: Fix parser.py for changed typing module
The typing module has subtle changes that are not even
documented: Typing types now have a __name__ attribute.
That confused the parser of the pyi generator because
suddenly stingizing

    Callable[..., Optional[str]]

resulted in

    Callable[..., Optional]

because of special rules that return the generic name
of a typing type, which was very unexpected. Finding this bug
took a lot of debugging of the recursive `_resolve_type`
function.

We finally move the debugging_aid string as a function into
lib/tool.py, because this was very helpful.

Some changes are not valid in 5.15, because there were other
modification in 6.x source code, like f-string, and some Qt install dir
struncture, like Qt/resources.

[ChangeLog][shiboken6] The parser for .pyi files needed an update
because of undocumented changes in typing.py for Python 3.10 .

Task-number: PYSIDE-1436
Change-Id: I3b8f2c1aa52a23014a8a915a0c677af96dfc536f
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 2530cb3f165ac02b8f7132e3f5ab4f7f6896dbd9)
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>