py3.10-prep: re-implement zip import
The functions used in the implementation of the embedded zip
importer will be deprecated in Python 3.10 .

[ChangeLog][shiboken6] The Zip importer was rewritten for the
Python 3.4 API which avoids warnings in Python 3.10 .

It also turned out that Python 3.10 gets recursion problems
when the imports are not resolved early. Since we always use
embedding, a late import in the functions does no longer save
anything.

This was actually a left-over to make the back-port to Python 2.7
easier, but this patch will no longer be back-ported.

Task-number: PYSIDE-1436
Change-Id: Ifc47e34951e36b42b92c60ee683b9a93c27d14c5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>