Fix crashes when using the QImage(uchar *) data constructors, take 2
The constructors expect a range of memory that remains valid
through the lifetime of the image and may also modify it.
Crashes occurred since apparently the Py_Buffer was released before
it.

To fix this, inject code to increase the reference count
and decrease it in the cleanup function.

Fixes: PYSIDE-1563
Change-Id: Ic4a47e0f458ed5b59b88813139102a9f1490a9df
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit 7079bc4ffc63dd42054014b4b78bff78abed962b)