shiboken6: Mark old syntax for conversion rules as deprecated
TypeEntry had a string member for conversion rules (besides
the nested "target-to-native" and "native-to-target" attributes)
that was populated from the "file" attribute depending on the "class"
attribute ("target"/"native"). Remove code path and flags for "native"
since they were not used. Rename the member to targetConversionRule()
for clarity and add a warning and FIXME comments.

Pick-to: 6.1
Change-Id: I2a991d438e48c1cc0519d077cb3c0599f9800eb7
Reviewed-by: Christian Tismer <tismer@stackless.com>