shiboken6: Add a check-function attribute to custom type entries
shiboken knows some check functions and defaults to
typeName + "_Check" in the end.

For custom types in modified function arguments, it is useful to be
able to specify a custom check function so that the overload decisor
finds the right type. Add a CustomTypeEntry class with a checkFunction
member for this.

[ChangeLog][shiboken6] It is now possible to specify a check function
for custom types in the typesystem.

Pick-to: 6.1
Change-Id: I21ec541afd5073dede09d359f2e2e16a0c6a0a05
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>